+++ styles/src/themes.ts
@@ -76,6 +76,9 @@ export interface Theme {
   state_label_halo: string;
   country_label: string;
 
+  address_label: string;
+  address_label_halo: string;
+
   regular?: string;
   bold?: string;
   italic?: string;
@@ -188,6 +191,9 @@ export const CONTRAST: Theme = {
   state_label: "#777777",
   state_label_halo: "#ffffff",
   country_label: "#9590aa",
+
+  address_label: "#91888b",
+  address_label_halo: "#ffffff",
 };
 
 export const LIGHT: Theme = {
@@ -269,6 +275,9 @@ export const LIGHT: Theme = {
   state_label: "#b3b3b3",
   state_label_halo: "#e0e0e0",
   country_label: "#a3a3a3",
+
+  address_label: "#91888b",
+  address_label_halo: "#ffffff",
 
   pois: {
     blue: "#1A8CBD",
@@ -372,6 +381,9 @@ export const DARK: Theme = {
   state_label_halo: "#1f1f1f",
   country_label: "#5c5c5c",
 
+  address_label: "#525252",
+  address_label_halo: "#1f1f1f",
+
   pois: {
     blue: "#4299BB",
     green: "#30C573",
@@ -473,6 +485,9 @@ export const WHITE: Theme = {
   state_label: "#b3b3b3",
   state_label_halo: "#ffffff",
   country_label: "#b8b8b8",
+
+  address_label: "#adadad",
+  address_label_halo: "#ffffff",
 };
 
 export const GRAYSCALE: Theme = {
@@ -554,6 +569,9 @@ export const GRAYSCALE: Theme = {
   state_label: "#999999",
   state_label_halo: "#cccccc",
   country_label: "#858585",
+
+  address_label: "#999999",
+  address_label_halo: "#e0e0e0",
 };
 
 export const BLACK: Theme = {
@@ -635,6 +653,9 @@ export const BLACK: Theme = {
   state_label: "#3d3d3d",
   state_label_halo: "#141414",
   country_label: "#707070",
+
+  address_label: "#525252",
+  address_label_halo: "#141414",
 };
 
 const themes: Record<string, Theme> = {