+++ tiles/src/main/java/com/protomaps/basemap/Basemap.java
@@ -94,7 +94,7 @@ public class Basemap extends ForwardingProfile {
 
   @Override
   public String version() {
-    return "4.2.0";
+    return "4.3.0";
   }
 
   @Override